feat(三交三问):三交三问答题
三交三问答题改为3题，且答错不再循环

+++ src/main/java/com/zzj/dto/req/ExamRecordReqDTO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 public class ExamRecordReqDTO {
 
-    private Integer reId;
+    private Long reId;
     private Long userId;
     private Double rePercent;
     private Integer reCount;
